v 0.3.7

+++ asserts/sourse/app.py
@@ -1,17 +1,19 @@
-from typing import Union, Sequence, List, Tuple, Optional, AnyStr as StrPath, Callable
+from enum import Enum
+from typing import Union, Sequence, List, Tuple, Optional, AnyStr as StrPath, Callable, Literal
+from json import load, dumps
 
 import pygame as p
 from pygame.math import Vector2
+from pygame.mouse import get_pos as get_cursor_pos
 from pygame.sprite import Sprite
-from pygame.mouse import get_pos as cursor_pos
-from pygame.font import FontType, get_default_font
 
 import asserts.graphics.graphics_manager as graphics
 import asserts.maps.maps_manager as maps
 import asserts.sourse.base_app as base_app
 import asserts.sourse.settings as settings
-from asserts.sounds.sounds_manager import Sounds, SoundPlayer
 from asserts.graphics.graphics_manager import AnimatedSprite, MultipleStateAnimatedSprite
+from asserts.sounds.sounds_manager import Sounds, SoundPlayer
+from asserts.sourse.button import Button
 
 
 # noinspection PyUnusedLocal,PyUnreachableCode
@@ -143,7 +145,7 @@ class Player(Sprite):
                     return ""
                 return n + ": " + str(var) + (" type: " + var.__class__.__name__ if t else "")
 
-            print_debug(f(self.pos, "pos", t=True), f(self.jumping, "jumping", r=True), f(cursor_pos, "cursor pos"),
+            print_debug(f(self.pos, "pos", t=True), f(self.jumping, "jumping", r=True), f(get_cursor_pos, "cursor pos"),
                         f(self.time, "time"))
         self.update()
         if self.is_dead:
@@ -235,42 +237,6 @@ class KilledPlayer:
         pass
 
 
-class Button:
-
-    def __init__(self, screen: p.Surface, rect: p.Rect = (0, 0, 0, 0),
-                 text: str = None, text_font: FontType = None, text_color: p.Color = None,
-                 text_bg_color: p.Color = None, img: Union[str, p.Surface, Sprite] = None, render=False):
-        self.rect = rect
-        self.screen = screen
-        if text is not None:
-            if text_font is None:
-                text_font = get_default_font
-            if text_color is None:
-                text_color = p.color.Color(255, 255, 255)
-            text = text_font.render(text, 1, text_color)
-            size = text.get_size()
-            self.display = p.Surface(size)
-            self.display.fill(text_bg_color)
-            self.display.blit(text, (0, 0))
-        elif img is not None:
-            if isinstance(img, str):
-                self.display = p.image.load(img)
-            elif isinstance(img, Sprite):
-                self.display = img.image
-            elif isinstance(img, p.Surface):
-                self.display = img
-            else:
-                raise ValueError
-
-        else:
-            raise ValueError
-        if render:
-            self.render()
-
-    def render(self):
-        self.screen.blit(self.display, self.rect.size)
-
-
 class GlobalizedSprites(p.sprite.Group):
     def __init__(self, *sprites: Union[Sprite, Sequence[Sprite]]):
         super().__init__(*sprites)
@@ -289,6 +255,7 @@ class GlobalizedSprites(p.sprite.Group):
         ret: List[p.Rect] = []
         for v in e:
             ret.append(v.hit_box)
+        return ret
 
 
 class EndGame(Exception):
@@ -303,12 +270,24 @@ class EndGame(Exception):
         super(EndGame, self).__init__("You win")
 
 
-class Level:
+class Screen:
+
+    def __init__(self, app):
+        self.app = app
+
+    def draw(self):
+        pass
+
+    def click(self, pos):
+        pass
+
+
+class Level(Screen):
     levels = range(3)
     levels_len = {1: 15, 2: 10, 3: 10}
 
     def __init__(self, level: int, app: "App"):
-        self.app = app
+        super(Level, self).__init__(app)
         if level not in self.levels:
             raise EndGame(True)
         le = maps.load_level(level)
@@ -393,18 +372,88 @@ class Level:
             dead_player.loop()
 
 
+class MainScreen(Screen):
+
+    def __init__(self, app):
+        super(MainScreen, self).__init__(app)
+
+        def click(s):
+            super(Button, s).click(s)
+            self.app.change_screen("level")
+
+        self.buttons = []
+        self.buttons.append(Button(self.app.screen, texts=["click, to run game"], action=click))
+
+    def draw(self):
+        for button in self.buttons:
+            button.render()
+
+    def click(self, cursor_pos):
+        for button in self.buttons:
+            button.check_click(cursor_pos)
+
+
+class SettingsScreen(Screen):
+
+    def __init__(self, app):
+        super(SettingsScreen, self).__init__(app)
+
+        def click(s):
+            super(Button, s).click(s)
+            self.app.change_screen("level")
+
+        self.buttons = []
+        self.buttons.append(Button(self.app.screen, (app.screen.get_width(), 0), texts=["back"], action=click))
+
+    def draw(self):
+        for button in self.buttons:
+            button.render()
+
+    def click(self, cursor_pos):
+        for button in self.buttons:
+            button.check_click(cursor_pos)
+
+    def draw(self):
+        pass
+
+
+class Screens(Enum):
+    main = (0, "m")
+    settings = (1, "m")
+    level1 = (1, "l")
+    level2 = (2, "l")
+    level3 = (3, "l")
+    level = "l"
+    levels = (level1, level2, level3, level)
+
+    def is_level(self):
+        return self.value in self.levels.value
+
+
 class App(base_app.BaseApp):
     def __init__(self, title="load again", icon_path: StrPath = "../graphics/icon.png", play_sound: bool = True,
                  height: int = 300, width: int = 300, bg_color: Tuple[int, int, int] = (0, 0, 0),
                  create_new_screen: bool = True):
         super().__init__(title, icon_path, height, width, bg_color, create_new_screen)
         self.level = Level(1, self)
+        self.settings = SettingsScreen(self)
+        self.main = MainScreen(self)
         self.player = Player(self.level.spawn.x, self.level.spawn.y,
                              graphics.get_player_sprite(settings.PLAYER_ANIMATION_TICKS, *self.level.spawn), self,
                              p.Rect(*self.level.spawn, *settings.PLAYER_SIZE))
         if play_sound:
             self.sound_player = SoundPlayer(Sounds.bgm, -1)
             self.sound_player.play()
+        self.sceneType = Screens.level1
+
+    @property
+    def scene(self) -> Screen:
+        if self.sceneType.is_level():
+            return self.level
+        elif self.sceneType == Screens.main:
+            return self.main
+        elif self.sceneType == Screens.settings:
+            return self.settings
 
     def on_key_pressed(self, key_code: int):
         if key_code == p.K_ESCAPE:
@@ -417,28 +466,49 @@ class App(base_app.BaseApp):
         #     self.player.right()
 
     def on_key_down(self, key_code: int):
-        if (key_code == p.K_SPACE) or (key_code == p.K_w) or (key_code == p.K_UP):
-            self.player.jump()
-        if (key_code == p.K_a) or (key_code == p.K_LEFT):
-            self.player.left()
-        if (key_code == p.K_d) or (key_code == p.K_RIGHT):
-            self.player.right()
-        if key_code == p.K_x:
-            self.player.kill()
+        if self.sceneType.is_level():
+            if key_code == p.K_ESCAPE:
+                self.change_screen("main")
+            else:
+                if (key_code == p.K_SPACE) or (key_code == p.K_w) or (key_code == p.K_UP):
+                    self.player.jump()
+                if (key_code == p.K_a) or (key_code == p.K_LEFT):
+                    self.player.left()
+                if (key_code == p.K_d) or (key_code == p.K_RIGHT):
+                    self.player.right()
+                if key_code == p.K_x:
+                    self.player.kill()
 
     def on_key_up(self, key_code: int):
         pass
 
+    def on_mouse_button_down(self, pos: Tuple[int, int], button_id: Literal[1, 2, 3, 4, 5]):
+        self.scene.click(pos)
+
+    def save(self):
+        pass
+
     def game_loop(self, delta):
-        try:
-            self.player.loop()
-            self.level.loop()
-        except EndGame as e:
-            if e.win:
-                self.level.next_level()
-            self.player.respawn()
-            self.level.respawn()
+        if self.sceneType.is_level():
+            try:
+                self.player.loop()
+                self.level.loop()
+            except EndGame as e:
+                if e.win:
+                    self.save()
+                    self.level.next_level()
+                self.player.respawn()
+                self.level.respawn()
 
     def draw(self):
-        self.player.draw()
-        self.level.draw()
+        if self.sceneType.is_level() or True:
+            self.player.draw()
+        self.scene.draw()
+
+    def change_screen(self, scene):
+        if isinstance(scene, str):
+            self.sceneType = Screens.__getattribute__(Screens, scene.lower())
+            if self.sceneType.value == "l":
+                self.sceneType = Screens.__getattribute__(Screens, "level" + str(self.level.level))
+        elif isinstance(scene, Screens):
+            self.sceneType = scene
